\name{amba}
\alias{amba}
\alias{is.amba}
\alias{amba.evaluate}
\alias{summary.amba}
\alias{fitted.amba}
\alias{fit.amba}
\alias{residuals.amba}
\title{additive models for business applications}
\description{todo}
\usage{
amba (y, fs, fit=TRUE, \dots)
is.amba (f)
amba.evaluate (m, us, \dots)
\method{summary}{amba}(m, which=NA, \dots)
\method{fitted}{amba}(m, \dots)
\method{fit}{amba}(m, \dots, nfits=12, preferential=TRUE)
\method{residuals}{amba}(m, which=NA, \dots)
}
\arguments{
\item{m}{.}
\item{fs}{.}
\item{which}{.}
\item{nfits}{.}
\item{preferential}{.}
\item{fit}{.}
\item{us}{.}
\item{f}{.}
\item{y}{.}
\item{\dots}{.}
}



