\name{categorical}
\alias{categorical}
\title{categorical terms}
\description{Creates categorical terms, noting that a categorical term is also a linear term. Unlike the lm function, there are estimates for each level of the variable, however an adjustment is made for summary output and inference.}
\usage{
categorical (x, y=NULL, valid.intercept=ifst (y), s)
}
\arguments{
\item{x}{}
\item{y}{}
\item{valid.intercept}{}
\item{s}{}
}

