\name{cmba}
\alias{cmba}
\alias{is.cmba}
\alias{print.cmba}
\alias{print.summary.cmba}
\title{contribution models for business applications}
\description{todo}
\usage{
cmba (y, fs)
is.cmba (f)
\method{print}{cmba}(m, \dots)
\method{print}{summary.cmba}(s, \dots)
}
\arguments{
\item{f}{.}
\item{fs}{.}
\item{m}{.}
\item{s}{.}
\item{y}{.}
\item{\dots}{.}
}

