\name{contribution}
\alias{contribution}
\alias{is.contribution}
\alias{print.contribution}
\alias{format.contribution}
\alias{summary.contribution}
\alias{print.summary.contribution}
\alias{fit.contribution}
\alias{fitted.contribution}
\alias{residuals.contribution}
\title{contributions}
\description{todo}
\usage{
contribution (x, name="x")
is.contribution (f)
\method{print}{contribution}(f, \dots)
\method{format}{contribution}(f, \dots)
\method{summary}{contribution}(f, y, \dots)
\method{print}{summary.contribution}(s, \dots)
\method{fit}{contribution}(f, y, raw=FALSE, \dots)
\method{fitted}{contribution}(f, \dots)
\method{residuals}{contribution}(f, y, \dots)
}
\arguments{
\item{f}{.}
\item{name}{.}
\item{raw}{.}
\item{x}{.}
\item{y}{.}
\item{s}{.}
\item{\dots}{.}
}




