\name{contributionset}
\alias{contributionset}
\alias{is.contributionset}
\alias{print.contributionset}
\alias{+.contribution}
\title{contribution sets}
\description{todo}
\usage{
contributionset (f)
is.contributionset (f)
\method{print}{contributionset}(m, \dots)
\method{+}{contribution}(a, b)
}
\arguments{
\item{f}{.}
\item{m}{.}
\item{a}{.}
\item{b}{.}
\item{\dots}{.}
}

