\name{eqnp}
\alias{eqnp}
\alias{eqnp.amba}
\alias{eqnp.linear}
\alias{eqnp.smooth}
\alias{eqnp.default}
\title{equivalent number of parameters}
\description{todo}
\usage{
eqnp (\dots)
\method{eqnp}{amba}(m, convergent.only=TRUE, \dots)
\method{eqnp}{linear}(t, \dots)
\method{eqnp}{smooth}(t, \dots)
\method{eqnp}{default}(\dots)
}
\arguments{
\item{m}{}
\item{convergent.only}{}
\item{t}{}
\item{\dots}{}
}



