\name{Estimation}
\alias{Estimation}
\alias{OrdinaryLeastSquares}
\alias{WeightedLeastSquares}
\alias{LocalPolynomialSmoothing}
\alias{is.estimation}
\alias{smooth.series}
\title{estimation container objects}
\description{todo}
\usage{
Estimation ()
OrdinaryLeastSquares (\dots)
WeightedLeastSquares (\dots)
LocalPolynomialSmoothing (\dots)
is.estimation (obj)
smooth.series (x, y, ns, sx, deg, wf, smoothness)
}
\arguments{
\item{obj}{.}
\item{x}{.}
\item{y}{.}
\item{ns}{.}
\item{sx}{.}
\item{deg}{.}
\item{wf}{.}
\item{smoothness}{.}
\item{\dots}{.}
}


