\name{evaluate.amba}
\alias{evaluate.amba}
\title{evaluate amba models}
\description{This function evaluates amba models in a similar way to evaluating terms. One slight complication is that the new x values, here u values, are a list of vectors (usually), one for each term.}
\usage{
\method{evaluate}{amba}(m, us, \dots)
}
\arguments{
\item{m}{}
\item{us}{}
\item{\dots}{}
}

