\name{gf}
\alias{gf.amba}
\alias{gf.term}
\alias{simple.gf}
\title{goodness of fit statistics for amba and term objects}
\description{Self-explanatory, except for the method argument, however these functions may be changed substantially in future.}
\usage{
\method{gf}{amba}(m, method="mar", \dots)
\method{gf}{term}(t, y, method="rsq", \dots)
simple.gf (yh, y, r=y-yh, method=c ("rsq", "mar") )
}
\arguments{
\item{m}{.}
\item{t}{.}
\item{yh}{.}
\item{y}{.}
\item{r}{.}
\item{method}{.}
\item{\dots}{.}
}


