\name{interaction}
\alias{interaction}
\alias{is.interaction}
\alias{interaction.predesign}
\alias{interaction.combine}
\alias{term.resist}
\alias{term.resists}
\title{interaction terms}
\description{Currently limited to two categorical variables. These have not yet been fully developed or tested.}
\usage{
interaction (t1, t2)
is.interaction (t)
interaction.predesign (x)
interaction.combine (x, op)
term.resist (t, tx)
term.resists (t, tx)
}
\arguments{
\item{t1}{.}
\item{t2}{.}
\item{x}{.}
\item{t}{.}
\item{tx}{.}
\item{op}{.}
}







