\name{linear.design}
\alias{linear.design}
\title{linear design objects}
\description{This function creates a linear design. The key argument is a character vector specifying function bodies. Symbolic manipulation is used to produce a list of actual R functions.}
\usage{
linear.design (key, valid.intercept, zptr=NA)
}
\arguments{
\item{key}{}
\item{valid.intercept}{}
\item{zptr}{}
}


