\name{linear.estimate}
\alias{linear.estimate}
\alias{print.linear.estimate}
\title{linear estimators and estimates}
\description{todo}
\usage{
linear.estimate (labs, th=rep (0, length (labs) ) )
\method{print}{linear.estimate}(e, \dots)
}
\arguments{
\item{labs}{Labels for each parameter estimate.}
\item{th}{Parameter estimates (theta hat).}
\item{e}{A linear estimate object.}
\item{\dots}{}
}


