\name{linear}
\alias{linear}
\alias{is.linear}
\alias{fitraw.linear}
\alias{summaryraw.linear}
\alias{fitted.linear}
\title{linear contributions}
\description{todo}
\usage{
linear (x, y, delegates=c ("1", "x"), name)
is.linear (f)
\method{fitraw}{linear}(f, y, \dots)
\method{summaryraw}{linear}(f, \dots)
\method{fitted}{linear}(f, \dots)
}
\arguments{
\item{f}{.}
\item{delegates}{.}
\item{name}{.}
\item{x}{.}
\item{y}{.}
\item{\dots}{.}
}

