\name{mmba}
\alias{mmba}
\alias{is.mmba}
\alias{fit.mmba}
\alias{summary.mmba}
\title{marginal models for business applications}
\description{todo}
\usage{
mmba (y, fs, fit=TRUE, \dots)
is.mmba (f)
\method{fit}{mmba}(m, \dots)
\method{summary}{mmba}(m, which=NA, \dots)
}
\arguments{
\item{f}{.}
\item{fs}{.}
\item{which}{.}
\item{fit}{.}
\item{m}{.}
\item{y}{.}
\item{\dots}{.}
}

