\name{mutate.term}
\alias{mutate.linear}
\alias{mutate.smooth}
\title{mutate methods for terms}
\description{These functions replace part of a term estimate object, with a raw estimate. What exactly that means depends on the term. Refer to fit.term for more information.}
\usage{
\method{mutate}{linear}(t, re, \dots)
\method{mutate}{smooth}(t, re, \dots)
}
\arguments{
\item{t}{}
\item{re}{}
\item{\dots}{}
}

