\name{polynomial}
\alias{polynomial}
\title{polynomial terms}
\description{Create a polynomial term object, noting the polynomial terms are linear terms, where linear refers to linearity in parameters, and polynomial refers to a function of x. A polynomial term adds little over the standard linear terms, other than taking care of key function specification.}
\usage{
polynomial (x, y=NULL, deg=1, valid.intercept=ifst (y), s)
}
\arguments{
\item{x}{.}
\item{y}{.}
\item{deg}{.}
\item{valid.intercept}{.}
\item{s}{.}
}




