\name{rawestimate}
\alias{rawestimate}
\alias{rawestimate.linear}
\alias{rawestimate.smooth}
\alias{rawestimate.default}
\title{extract raw estimate}
\description{These functions extract a raw estimate from a term object, that is, the object returned by an estimator.}
\usage{
rawestimate (\dots)
\method{rawestimate}{linear}(t, \dots)
\method{rawestimate}{smooth}(t, \dots)
\method{rawestimate}{default}(\dots)
}
\arguments{
\item{t}{}
\item{\dots}{}
}

