\name{reset}
\alias{reset.linear}
\alias{reset.smooth}
\title{reset term estimates}
\description{These function reset term estimates to their defaults, usually something that evaluates to zeros, noting that terms must have estimates, regardless of whether they have been fitted or not.}
\usage{
\method{reset}{linear}(t, \dots)
\method{reset}{smooth}(t, \dots)
}
\arguments{
\item{t}{.}
\item{\dots}{.}
}


