\name{sample}
\docType{data}
\alias{sample}
\title{simulated dataset}
\description{Simulated dataset for testing amba models.}
\format{
A table containing 102 realisations and 7 variables.
\tabular{rll}{
[,1] \tab g1 \tab factor <defective> \cr
[,2] \tab g2 \tab factor <defective> \cr
[,3] \tab x1 \tab numeric <defective> \cr
[,4] \tab x2  \tab numeric <defective> \cr
[,5] \tab x3 \tab numeric <defective> \cr
[,6] \tab x4 \tab numeric <defective>\cr
[,7] \tab y  \tab numeric <clean>
}}


