\name{seq.term}
\alias{seq.term}
\title{seq method for term objects}
\description{For term objects where nc = 1, produces a regularly spaced sequence from the minimum to the maximum realised value.}
\usage{
\method{seq}{term}(t, n=200, \dots)
}
\arguments{
\item{t}{.}
\item{n}{.}
\item{\dots}{.}
}

