\name{setestimation}
\alias{setestimation}
\alias{estimation.validate}
\alias{standardestimation}
\alias{standardestimation.linear}
\alias{standardestimation.smooth}
\alias{standardestimation.default}
\title{set estimators and inferencers}
\description{todo}
\usage{
setestimation (ts, ec, \dots)
estimation.validate (t)
standardestimation (\dots)
\method{standardestimation}{linear}(t, \dots)
\method{standardestimation}{smooth}(t, \dots)
\method{standardestimation}{default}(\dots)
}
\arguments{
\item{t}{.}
\item{ts}{.}
\item{ec}{.}
\item{\dots}{.}
}


