\name{smooth.design}
\alias{smooth.design}
\title{create a smooth design}
\description{Refer to smooth, for information about these arguments.}
\usage{
smooth.design (ns, nfits, deg, wf, smoothness, x)
}
\arguments{
\item{ns}{}
\item{nfits}{}
\item{deg}{}
\item{wf}{}
\item{smoothness}{}
\item{x}{}
}

