\name{smooth.estimate}
\alias{smooth.estimate}
\alias{print.smooth.estimate}
\title{smooth estimators and estimates}
\description{todo}
\usage{
smooth.estimate (ns, th=0, sy=rep (0, ns) )
\method{print}{smooth.estimate}(e, \dots)
}
\arguments{
\item{ns}{}
\item{th}{}
\item{sy}{}
\item{e}{}
\item{\dots}{}
}

