\name{summary.term}
\alias{summary.term}
\alias{summary.linear}
\alias{summary.smooth}
\alias{print.summary.term}
\alias{print.summaryext.smooth}
\title{summary output for terms}
\description{todo}
\usage{
\method{summary}{term}(t, y, np=eqnp (t), \dots)
\method{summary}{linear}(t, \dots)
\method{summary}{smooth}(t, \dots)
\method{print}{summary.term}(s, \dots)
\method{print}{summaryext.smooth}(s, \dots)
}
\arguments{
\item{t}{}
\item{y}{}
\item{np}{}
\item{s}{}
\item{\dots}{}
}



