\name{term.pump}
\alias{term.pump}
\title{support function for term constructors}
\description{This function should not be called by users. It is intended to be called by term subclass constructors. If y is provided it fits the term. The sc argument is used to extend the term class attributed.}
\usage{
term.pump (t, y, sc=NULL)
}
\arguments{
\item{t}{.}
\item{y}{.}
\item{sc}{.}
}


