\name{term.validate}
\alias{term.validate}
\title{validate term objects}
\description{This function is not intended to be called by users, and is called inside the term superclass constructor. It ensures that the term attributes nr, nc, nv and valid, are correct.}
\usage{
term.validate (t)
}
\arguments{
\item{t}{.}
}


