\name{termlist}
\alias{termlist}
\alias{is.termlist}
\alias{as.termlist}
\alias{print.termlist}
\alias{+.term}
\alias{-.term}
\title{termlist objects}
\description{todo}
\usage{
termlist ()
is.termlist (ts)
as.termlist (obj, validate=TRUE)
\method{print}{termlist}(ts, \dots)
\method{+}{term}(t1, t2)
\method{-}{term}(t1, t2)
}
\arguments{
\item{obj}{.}
\item{validate}{.}
\item{ts}{.}
\item{t1}{.}
\item{t2}{.}
\item{\dots}{.}
}


