\name{validex}
\alias{validex}
\title{explicit valid vector}
\description{For clean terms, the valid attribute is null, for unclean terms it is a logical vector. This function produces a valid logical vector, regardless of whether the term is clean or not.}
\usage{
validex (t)
}
\arguments{
\item{t}{.}
}


