\name{categorical}
\alias{categorical}
\alias{is.categorical}
\alias{fitraw.categorical}
\alias{fitted.categorical}
\title{categorical contributions}
\description{todo}
\usage{
categorical (x, y, name)
is.categorical (f)
\method{fitraw}{categorical}(f, y, \dots)
\method{fitted}{categorical}(f, \dots)
}
\arguments{
\item{f}{.}
\item{name}{.}
\item{x}{.}
\item{y}{.}
\item{\dots}{.}
}

