\name{mmba}
\alias{mmba}
\alias{fit.mmba}
\alias{plot.mmba}
\title{marginal models for business applications}
\description{This a very simple model, that fits each term completely seperately, onto the response. It's syntax is very similar to amba.}
\usage{
mmba (y, ts, fit=TRUE, \dots)
\method{fit}{mmba}(m, \dots)
\method{plot}{mmba}(m, ..., index=FALSE)
}
\arguments{
\item{y}{}
\item{ts}{}
\item{fit}{}
\item{m}{}
\item{index}{}
\item{\dots}{}
}


