\name{plots}
\alias{plot.mmba}
\alias{plot.amba}
\alias{pairs.cmba}
\alias{pairs.contributionset}
\alias{plot.contribution}
\alias{plot.categorical}
\alias{lines.contribution}
\title{plots}
\description{todo}
\usage{
\method{plot}{mmba}(m, which=NA, index=FALSE, \dots)
\method{plot}{amba}(m, which=NA, index=FALSE, \dots)
\method{pairs}{cmba}(m, \dots)
\method{pairs}{contributionset}(fs, \dots)
\method{plot}{contribution}(f, y, index=FALSE, main, xlab, ylab, \dots)
\method{plot}{categorical}(f, y, index=FALSE, main, xlab, ylab, \dots)
\method{lines}{contribution}(f, \dots)
}
\arguments{
\item{m}{.}
\item{f}{.}
\item{fs}{.}
\item{which}{.}
\item{index}{.}
\item{main}{.}
\item{xlab}{.}
\item{ylab}{.}
\item{y}{.}
\item{\dots}{.}
}

