\name{sample}
\docType{data}
\alias{sample}
\title{simulated dataset}
\description{Simulated dataset for testing amba models.}
\format{
A table containing 102 rows (realisations) and 7 columns (variables). This data contains a lot of missing values.
\tabular{rll}{
[,1] \tab g1 \tab factor \cr
[,2] \tab g2 \tab factor \cr
[,3] \tab x1 \tab numeric \cr
[,4] \tab x2  \tab numeric \cr
[,5] \tab x3 \tab numeric \cr
[,6] \tab x4 \tab numeric \cr
[,7] \tab y  \tab numeric
}}

