\name{smooth}
\alias{smooth}
\alias{is.smooth}
\alias{fitraw.smooth}
\alias{summaryraw.smooth}
\title{smooth contributions}
\description{todo}
\usage{
smooth (x, y, ns=10, degree=2, smoothness=0.67, bw, name)
is.smooth (f)
\method{fitraw}{smooth}(f, y, \dots)
\method{summaryraw}{smooth}(f, \dots)
}
\arguments{
\item{f}{.}
\item{degree}{.}
\item{smoothness}{.}
\item{bw}{.}
\item{name}{.}
\item{ns}{.}
\item{x}{.}
\item{y}{.}
\item{\dots}{.}
}


